log: fix regression to "--source" when "--decorate" was updated
Split out the logic to show "--source" from showing "--decorate"
annotation into a separate show_source() helper.  Most of the time
this is called immediately before show_decorations(), but the oneline
output format is a special case that requires us to be able to show
the decorations alone at the very end while showing the source before
the message.

This allows us to revert the change to t4202 in the previous step,
which expected the regressed output for "--source".

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>