Revert "parse-remote: remove reference to unused op_prep"
This reverts commit 1a7e780f0dd7bea62aedd7bb57cd1df70bb4c3bd,
because we have no way of knowing how many third-party scripts
dot-source this scriptlet and rely on the current behaviour.

Without first breaking them, that is.  Breaking them first and
having to deal with fallout is not worth our effort.  The only thing
leaving this otherwise unused parameter in may hurt is the
performance, which does not matter in this error codepath.